Separate scheduler integrated tests from test_main into scheduler/tests/integrated/
