+++ requirements-dev.txt
@@ -3,6 +3,7 @@ requests>=2.4.0
 pyarrow>=0.11.0
 cython>=0.29
 mock>=2.0.0; python_version < '3'
+flaky>=3.5.0
 pytest>=3.5.0
 pytest-cov>=2.5.0
 pytest-timeout>=1.2.0
